requirements.txt
amqp==2.5.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.2.3
django-filter==2.1.0
django-filters==0.2.1
djangorestframework==3.9.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.20
kombu==4.6.3
lazy-object-proxy==1.4.1
lxml==4.3.4
mccabe==0.6.1
parsel==1.5.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pytz==2019.1
queuelib==1.5.0
redis==3.2.1
requests==2.22.0
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
scrapy-proxies==0.4
Scrapy-UserAgents==0.0.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
Twisted==19.2.1
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
w3lib==1.20.0
wrapt==1.11.2
zope.interface==4.6.0
